[refactor](publish) Move some publishing endpoints to PublishModule
