+++ composeApp/src/commonMain/kotlin/world/hachimi/app/api/module/SongModule.kt
@@ -1,11 +1,5 @@
 package world.hachimi.app.api.module
 
-import io.ktor.client.content.*
-import io.ktor.client.plugins.*
-import io.ktor.client.request.*
-import io.ktor.client.request.forms.*
-import io.ktor.http.*
-import kotlinx.io.Source
 import kotlinx.serialization.Serializable
 import world.hachimi.app.api.ApiClient
 import world.hachimi.app.api.WebResult
@@ -142,118 +136,6 @@ class SongModule(
 
     suspend fun detailById(id: Long): WebResult<PublicSongDetail> =
         client.get("/song/detail_by_id", DetailByIdReq(id), false)
-
-    @Serializable
-    data class UploadImageResp(
-        val tempId: String
-    )
-
-    suspend fun uploadCoverImage(
-        filename: String,
-        source: Source,
-        listener: ProgressListener? = null
-    ): WebResult<UploadImageResp> {
-        return client.postWith("/song/upload_cover_image") {
-            setBody(
-                MultiPartFormDataContent(
-                    formData {
-                        append(
-                            "image",
-                            InputProvider { source },
-                            headersOf(HttpHeaders.ContentDisposition, "filename=\"${filename}\"")
-                        )
-                    }
-                )
-            )
-            onUpload(listener)
-        }
-    }
-
-    @Serializable
-    data class UploadAudioFileResp(
-        val tempId: String,
-        val durationSecs: Long,
-        val title: String?,
-        val bitrate: String?,
-        val artist: String?,
-    )
-
-    suspend fun uploadAudioFile(
-        filename: String,
-        source: Source,
-        listener: ProgressListener? = null
-    ): WebResult<UploadAudioFileResp> {
-        return client.postWith("/song/upload_audio_file") {
-            setBody(
-                MultiPartFormDataContent(
-                    formData {
-                        append(
-                            "audio",
-                            InputProvider { source },
-                            headersOf(HttpHeaders.ContentDisposition, "filename=\"${filename}\"")
-                        )
-                    }
-                )
-            )
-            timeout {
-                socketTimeoutMillis = 60_000
-                connectTimeoutMillis = 60_000
-                requestTimeoutMillis = 60_000
-            }
-            onUpload(listener)
-        }
-    }
-
-
-    @Serializable
-    data class PublishReq(
-        val songTempId: String,
-        val coverTempId: String,
-        val title: String,
-        val subtitle: String,
-        val description: String,
-        val lyrics: String,
-        val tagIds: List<Long>,
-        val creationInfo: CreationInfo,
-        val productionCrew: List<ProductionItem>,
-        val externalLinks: List<ExternalLink>,
-        /**
-         * @since 251105
-         */
-        val explicit: Boolean?,
-        /**
-         * @since 251114, should be required in new client.
-         */
-        val jmid: String?,
-        /**
-         * @since 251114
-         */
-        val comment: String?,
-    ) {
-        @Serializable
-        data class CreationInfo(
-            // 0: original, 1: derivative work, 2: tertiary work
-            val creationType: Int,
-            val originInfo: CreationTypeInfo?,
-            val derivativeInfo: CreationTypeInfo?,
-        )
-
-        @Serializable
-        data class ProductionItem(
-            val role: String,
-            val uid: Long?,
-            val name: String?,
-        )
-    }
-
-    @Serializable
-    data class PublishResp(
-        val songDisplayId: String
-    )
-
-    suspend fun publish(req: PublishReq): WebResult<PublishResp> =
-        client.post("/song/publish", req)
-
     @Serializable
     data class SearchReq(
         val q: String,
